use NSDocumentController for opening movie files
